Add tests and separate SearchForm component
