+++ packages/catalog-frontend/package.json
@@ -36,5 +36,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "src/{**/,}{!(index),}.tsx"
+    ]
   }
 }
