--- packages/catalog-frontend/src/App.test.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import App from './App';
-
-test('renders learn react link', () => {
-  const { getByText } = render(<App />);
-  const linkElement = getByText(/CDK Construct Catalog/i);
-  expect(linkElement).toBeInTheDocument();
-});
